Catalyst App/functions/catalyst_backend/requirements.txt
# Zoho Catalyst SDK
zcatalyst-sdk==1.2.0

# Flask framework
Flask==2.3.2
Flask-Cors==3.0.10
Werkzeug>=2.3.0

# HTTP requests
requests>=2.28.0

# Environment variables
python-dotenv>=1.0.0

# Authentication
PyJWT==2.8.0

# Password hashing (optional - has fallback to SHA-256)
bcrypt==4.1.2

# Caching
cachetools==5.3.2
